requirements.txt
apache-airflow==2.2.4
requests==2.27.1
pandas==1.4.2
psycopg2-binary==2.9.3